# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 孝感非遗文化数字化平台 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type HeritageItemLevel =
  (typeof HeritageItemLevel)[keyof typeof HeritageItemLevel];

export const HeritageItemLevel = {
  national: "national",
  provincial: "provincial",
  municipal: "municipal",
} as const;

export interface HeritageItem {
  id: number;
  name: string;
  nameEn?: string;
  level: HeritageItemLevel;
  category: string;
  description: string;
  shortDesc: string;
  imageUrl: string;
  videoUrl?: string;
  origin?: string;
  yearListed?: number;
  artisanCount?: number;
  tags?: string[];
  featured?: boolean;
  /** 是否与孝文化直接相关 */
  xiaoTheme?: boolean;
}

export interface HeritageItemListResponse {
  items: HeritageItem[];
  total: number;
  page: number;
  limit: number;
}

export interface Category {
  id: string;
  name: string;
  count: number;
  icon: string;
}

export interface CategoryListResponse {
  categories: Category[];
}

export type ArtisanServiceType =
  (typeof ArtisanServiceType)[keyof typeof ArtisanServiceType];

export const ArtisanServiceType = {
  workshop: "workshop",
  custom: "custom",
  course: "course",
} as const;

export interface ArtisanService {
  id: number;
  type: ArtisanServiceType;
  title: string;
  description?: string;
  price: number;
  duration?: string;
  maxParticipants?: number;
}

export interface Artisan {
  id: number;
  name: string;
  level: string;
  heritageItem: string;
  heritageItemId?: number;
  bio: string;
  avatarUrl: string;
  yearsOfPractice?: number;
  awards?: string[];
  services?: ArtisanService[];
}

export interface ArtisanListResponse {
  artisans: Artisan[];
  total: number;
}

export type QuizQuestionDifficulty =
  (typeof QuizQuestionDifficulty)[keyof typeof QuizQuestionDifficulty];

export const QuizQuestionDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  expert: "expert",
} as const;

export interface QuizQuestion {
  id: number;
  question: string;
  options: string[];
  difficulty: QuizQuestionDifficulty;
  category?: string;
  points: number;
}

export interface QuizQuestionListResponse {
  questions: QuizQuestion[];
}

export interface QuizSubmitRequest {
  questionId: number;
  selectedAnswer: number;
}

export interface QuizSubmitResponse {
  correct: boolean;
  correctAnswer: number;
  explanation: string;
  pointsEarned: number;
  totalPoints: number;
}

export type ProductCategory =
  (typeof ProductCategory)[keyof typeof ProductCategory];

export const ProductCategory = {
  artisan_original: "artisan_original",
  collab: "collab",
  student: "student",
} as const;

export interface Product {
  id: number;
  name: string;
  category: ProductCategory;
  description: string;
  price: number;
  imageUrl: string;
  artisanName?: string;
  heritageItem?: string;
  stock?: number;
  rating?: number;
  reviewCount?: number;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
}

export type ActivityStatus =
  (typeof ActivityStatus)[keyof typeof ActivityStatus];

export const ActivityStatus = {
  upcoming: "upcoming",
  ongoing: "ongoing",
  past: "past",
} as const;

export interface Activity {
  id: number;
  title: string;
  description: string;
  type: string;
  location: string;
  startDate: string;
  endDate?: string;
  imageUrl?: string;
  status: ActivityStatus;
  registrationUrl?: string;
  maxParticipants?: number;
  currentParticipants?: number;
}

export interface ActivityListResponse {
  activities: Activity[];
}

/**
 * 生成风格
 */
export type AiGenerateRequestStyle =
  (typeof AiGenerateRequestStyle)[keyof typeof AiGenerateRequestStyle];

export const AiGenerateRequestStyle = {
  papercut: "papercut",
  shadow_puppet: "shadow_puppet",
  xiao_culture: "xiao_culture",
  plaster_carving: "plaster_carving",
  kiln: "kiln",
} as const;

/**
 * 应用场景
 */
export type AiGenerateRequestScene =
  (typeof AiGenerateRequestScene)[keyof typeof AiGenerateRequestScene];

export const AiGenerateRequestScene = {
  poster: "poster",
  phone_case: "phone_case",
  bookmark: "bookmark",
  packaging: "packaging",
  avatar: "avatar",
  greeting_card: "greeting_card",
} as const;

export interface AiGenerateRequest {
  /** 生成风格 */
  style: AiGenerateRequestStyle;
  /** 应用场景 */
  scene: AiGenerateRequestScene;
  /** 用户输入的描述 */
  prompt: string;
}

export interface AiGenerateResponse {
  imageUrl: string;
  style: string;
  scene: string;
  prompt: string;
  generatedAt: string;
}

export interface PlatformStats {
  nationalItems: number;
  provincialItems: number;
  municipalItems: number;
  artisanCount: number;
  productCount: number;
  visitorCount: number;
  cityName: string;
}

export type ListHeritageItemsParams = {
  level?: ListHeritageItemsLevel;
  category?: string;
  page?: number;
  limit?: number;
};

export type ListHeritageItemsLevel =
  (typeof ListHeritageItemsLevel)[keyof typeof ListHeritageItemsLevel];

export const ListHeritageItemsLevel = {
  national: "national",
  provincial: "provincial",
  municipal: "municipal",
  all: "all",
} as const;

export type ListArtisansParams = {
  level?: string;
  page?: number;
  limit?: number;
};

export type ListQuizQuestionsParams = {
  difficulty?: ListQuizQuestionsDifficulty;
  limit?: number;
};

export type ListQuizQuestionsDifficulty =
  (typeof ListQuizQuestionsDifficulty)[keyof typeof ListQuizQuestionsDifficulty];

export const ListQuizQuestionsDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  expert: "expert",
} as const;

export type ListProductsParams = {
  category?: string;
  page?: number;
  limit?: number;
};

export type ListActivitiesParams = {
  status?: ListActivitiesStatus;
};

export type ListActivitiesStatus =
  (typeof ListActivitiesStatus)[keyof typeof ListActivitiesStatus];

export const ListActivitiesStatus = {
  upcoming: "upcoming",
  ongoing: "ongoing",
  past: "past",
} as const;