# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 孝感非遗文化数字化平台 API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityStatus } from "./activityStatus";

export interface Activity {
  id: number;
  title: string;
  description: string;
  type: string;
  location: string;
  startDate: string;
  endDate?: string;
  imageUrl?: string;
  status: ActivityStatus;
  registrationUrl?: string;
  maxParticipants?: number;
  currentParticipants?: number;
}