# lib/api-zod/src/generated/types/categoryListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 孝感非遗文化数字化平台 API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from "./category";

export interface CategoryListResponse {
  categories: Category[];
}