# lib/api-zod/src/generated/types/heritageItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 孝感非遗文化数字化平台 API
 * OpenAPI spec version: 0.1.0
 */
import type { HeritageItemLevel } from "./heritageItemLevel";

export interface HeritageItem {
  id: number;
  name: string;
  nameEn?: string;
  level: HeritageItemLevel;
  category: string;
  description: string;
  shortDesc: string;
  imageUrl: string;
  videoUrl?: string;
  origin?: string;
  yearListed?: number;
  artisanCount?: number;
  tags?: string[];
  featured?: boolean;
  /** 是否与孝文化直接相关 */
  xiaoTheme?: boolean;
}