# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 孝感非遗文化数字化平台 API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductCategory } from "./productCategory";

export interface Product {
  id: number;
  name: string;
  category: ProductCategory;
  description: string;
  price: number;
  imageUrl: string;
  artisanName?: string;
  heritageItem?: string;
  stock?: number;
  rating?: number;
  reviewCount?: number;
}