# lib/api-zod/src/generated/types/quizSubmitResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 孝感非遗文化数字化平台 API
 * OpenAPI spec version: 0.1.0
 */

export interface QuizSubmitResponse {
  correct: boolean;
  correctAnswer: number;
  explanation: string;
  pointsEarned: number;
  totalPoints: number;
}